patches: Add patches to build a Raspberry Pi 4 image from RPi upstream kernel
Add a set of diff patches that are applied on top of the RPi5 patches (that are
in turn applied on top of the upstream Raspberry Pi kernel) in order to build a
RPi 4 image.
Update the Makefile to use the patches and to be able to build the two types of
image.