Makefile,patches: Add build caching for kernel pkgs build
Add `cache-to` and `cache-from` instructions for the kernel pkgs build. From
testing this is almost all of the build time. The final image is one single
layer so the cache is too, this may prove to be inefficient when it comes to
building new kernels but it should allow efficient repeats when running the
Makefile.

AUTO-15741