Makefile: Produce consistently tagged production builds for AutInf
Change the image tagging system so that instead of using git repo based tags
all container images are tagged with a constant tag that we can set at build
time.
There are limitations on the tag as described in the README.

We do not want this repo to build the actual `installer`, that is done in
`talos-k8s` repo with some extra arguments to define system extensions and
output type.

AUTO-15739